+++ mayan/apps/documents/settings.py
@@ -83,11 +83,17 @@ setting_print_height = namespace.add_setting(
 setting_print_width = namespace.add_setting(
     global_name='DOCUMENTS_PRINT_WIDTH', default='3600'
 )
-setting_recent_count = namespace.add_setting(
-    global_name='DOCUMENTS_RECENT_COUNT', default=40,
+setting_recent_access_count = namespace.add_setting(
+    global_name='DOCUMENTS_RECENT_ACCESS_COUNT', default=40,
     help_text=_(
-        'Maximum number of recent (created, edited, viewed) documents to '
-        'remember per user.'
+        'Maximum number of recently accessed (created, edited, viewed) '
+        'documents to remember per user.'
+    )
+)
+setting_recent_added_count = namespace.add_setting(
+    global_name='DOCUMENTS_RECENT_ADDED_COUNT', default=40,
+    help_text=_(
+        'Maximum number of recently created documents to show.'
     )
 )
 setting_rotation_step = namespace.add_setting(